refactor: move pagination logic to useLoadItems hook to improve reusability
